refactor: wrap geoLocation logic and move result setting into useQuery callback
