+++ apps/web/core/system-ids.ts
@@ -1,4 +1,4 @@
-import { Id } from '@graphprotocol/grc-20'
+import { Id } from '@graphprotocol/grc-20';
 
 // TODO: Temporary place for system IDs until we can use updated grc-20 package
 
@@ -6,6 +6,6 @@ export const PLACE_TYPE = Id.Id('Fr887xssrH7RbK3S5gnLVb');
 export const SOURCES_TYPE = Id.Id('A7NJF2WPh8VhmvbfVWiyLo');
 
 export const VENUE_PROPERTY = Id.Id('Wx8o6Dahq5v3HhVSaLgwXn');
-export const ADDRESS_PROPERTY = Id.Id('FAgoYRgSim3ydKxzt5CDr5');
+export const ADDRESS_PROPERTY = Id.Id('HXLpAZyQkcy6Di4YJu4xzU');
 export const MAPBOX_PROPERTY = Id.Id('XrRygNVtYLwG4S1YoiWW8c');
 export const SOURCE_DATABASE_IDENTIFIER_PROPERTY = Id.Id('CgLt3CoEzWmhPW3XGkakYa');
