+++ apps/web/partials/entity-page/editable-entity-page.tsx
@@ -10,7 +10,6 @@ import * as React from 'react';
 import { StoreRelation } from '~/core/database/types';
 import { DB } from '~/core/database/write';
 import { useEditEvents } from '~/core/events/edit-events';
-import { useGeoCoordinates } from '~/core/hooks/use-geo-coordinates';
 import { useProperties } from '~/core/hooks/use-properties';
 import { useRelationship } from '~/core/hooks/use-relationship';
 import { useRenderables } from '~/core/hooks/use-renderables';
@@ -18,6 +17,7 @@ import { ID } from '~/core/id';
 import { EntityId } from '~/core/io/schema';
 import { useEditorStore } from '~/core/state/editor/use-editor';
 import { useEntityPageStore } from '~/core/state/entity-page-store/entity-store';
+import { VENUE_PROPERTY } from '~/core/system-ids';
 import {
   PropertySchema,
   Relation,
@@ -34,7 +34,7 @@ import { Checkbox, getChecked } from '~/design-system/checkbox';
 import { LinkableRelationChip } from '~/design-system/chip';
 import { DateField } from '~/design-system/editable-fields/date-field';
 import { ImageZoom, PageImageField, PageStringField } from '~/design-system/editable-fields/editable-fields';
-import { GeoLocationPointFields } from '~/design-system/editable-fields/geo-location-field';
+import { GeoLocationPointFields, GeoLocationWrapper } from '~/design-system/editable-fields/geo-location-field';
 import { NumberField } from '~/design-system/editable-fields/number-field';
 import { WebUrlField } from '~/design-system/editable-fields/web-url-field';
 import { Create } from '~/design-system/icons/create';
@@ -50,7 +50,6 @@ import { getRenderableTypeSelectorOptions } from './get-renderable-type-options'
 import { NumberOptionsDropdown } from './number-options-dropdown';
 import { RenderableTypeDropdown } from './renderable-type-dropdown';
 import { editorHasContentAtom } from '~/atoms';
-import { VENUE_PROPERTY } from '~/core/system-ids';
 
 interface Props {
   triples: ITriple[];
@@ -341,8 +340,6 @@ export function RelationsGroup({ relations, properties }: RelationsGroupProps) {
   const relationValueTypes = property?.relationValueTypes;
   const hasPlaceholders = relations.some(r => r.placeholder === true);
   const valueType = relationValueTypes?.[0];
-
-  const geoData = useGeoCoordinates(id, spaceId);
 
   return (
     <div className="flex flex-wrap items-center gap-1 pr-10">
@@ -680,8 +677,7 @@ export function RelationsGroup({ relations, properties }: RelationsGroupProps) {
             <div
               key={`relation-${relationId}-${relationValue}`}
               className={`mt-1 ${
-                renderableType === 'PLACE' ||
-                (renderableType === 'RELATION' && r.attributeId === VENUE_PROPERTY)
+                renderableType === 'PLACE' || (renderableType === 'RELATION' && r.attributeId === VENUE_PROPERTY)
                   ? 'w-full'
                   : ''
               }`}
@@ -706,18 +702,7 @@ export function RelationsGroup({ relations, properties }: RelationsGroupProps) {
               // the entity ends up with type = 'RELATION' after creation.
               // So temporary I'll add some checks to render it
               (renderableType === 'RELATION' && r.attributeId === VENUE_PROPERTY) ? (
-                <div className="flex w-full flex-col">
-                  <span className="my-3 text-[19px] leading-[29px]">{geoData?.name}</span>
-                  <GeoLocationPointFields
-                    key={relationId}
-                    variant="body"
-                    placeholder="Add value..."
-                    aria-label="text-field"
-                    value={geoData?.geoLocation}
-                    onChange={() => {}}
-                    hideInputs={true}
-                  />
-                </div>
+                <GeoLocationWrapper relationId={relationId} />
               ) : null}
             </div>
           );
